Update file structures
